requirements.txt
django==5.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-tenants==3.7.0
psycopg2-binary==2.9.9
gunicorn==23.0.0
dj-database-url==2.2.0
openpyxl==3.1.2